Native Language Server integration with PM
The first attempt at
a) building `enso` native-image that includes full Language Server
b) integrating the executable as an experimental feature during project
startup

The change (for now) assumes that `enso` executable appears in Enso's
default `engines` directory.
To build and run the new integration one has to
a) `engine-runner/buildNativeImage`
b) run PM with `--native-language-server`

This change also adds a copy of some of logback's code (`SocketAppender`
or a simple socket server`) as it was impossible to debug serialization
bugs without some additional logging.