Disable visualizations for subexpressions
The possibility of attaching visualizations to subexpressions meant that
we (currently) have to invalidate caches for their parent expression
every time a request comes. That was an acceptable cost when an
expression is relatively fast to compute but unacceptable when dealing
with slow computations that would have to be repeated.
Since currently attaching visualizations is not used at all and we can
rely on caching RHS and self, it is _safe_ to disable it. An observable
pattern is better suited for visualizations and would mitigate this
problem by design.